requirements.txt
python-dotenv==1.0.0
selenium==4.15.2
Flask==2.3.2
Flask-Cors==4.0.0
requests==2.31.0
openai==1.3.7
Pillow==10.1.0
numpy==1.26.2